controllers | move AppModel out to Application level & rename AppModel -> WhatToEatModel
