+++ src/main/java/david/buzas/whattoeat/WhatToEatApplication.java
@@ -1,6 +1,5 @@
 package david.buzas.whattoeat;
 
-import david.buzas.whattoeat.controllers.AppModel;
 import david.buzas.whattoeat.entities.Meal;
 import david.buzas.whattoeat.entities.MealCategory;
 import david.buzas.whattoeat.entities.MealConsumption;
@@ -19,7 +18,7 @@ public class WhatToEatApplication extends Application {
     public static Repository<MealCategory> mealCategoryRepository = new MealCategoryRepository("meal-categories.json");
     public static Repository<MealType> mealTypeRepository = new MealTypeRepository("meal-types.json");
 
-    public static AppModel model = new AppModel();
+    public static WhatToEatModel model = new WhatToEatModel();
 
     @Override
     public void start(Stage stage) throws IOException {
